organize code for garden lattice focused work
