+++ src/almanack/metrics/data.py
@@ -5,22 +5,20 @@ This module computes data for GitHub Repositories
 import json
 import logging
 import pathlib
-import re
 import shutil
 import tempfile
-import time
 from datetime import datetime, timedelta, timezone
 from typing import Any, Dict, List, Optional, Tuple
 from urllib.parse import urlparse
 
 import defusedxml.ElementTree as ET
 import pygit2
-import requests
 import yaml
 
 from ..git import (
     clone_repository,
     count_files,
+    file_exists_in_repo,
     find_file,
     get_commits,
     get_edited_files,
@@ -31,6 +29,15 @@ from .entropy.calculate_entropy import (
     calculate_aggregate_entropy,
     calculate_normalized_entropy,
 )
+from .garden_lattice.connectedness import (
+    count_unique_contributors,
+    default_branch_is_not_master,
+    detect_social_media_links,
+    find_doi_citation_data,
+)
+from .garden_lattice.practicality import count_repo_tags, get_ecosystems_package_metrics
+from .garden_lattice.understanding import includes_common_docs, is_citable
+from .remote import get_api_data
 
 LOGGER = logging.getLogger(__name__)
 
@@ -85,134 +92,6 @@ def get_table(repo_path: str) -> List[Dict[str, Any]]:
     ]
 
 
-def file_exists_in_repo(
-    repo: pygit2.Repository,
-    expected_file_name: str,
-    check_extension: bool = False,
-    extensions: list[str] = [".md", ""],
-) -> bool:
-    """
-    Check if a file (case-insensitive and with optional extensions)
-    exists in the latest commit of the repository.
-
-    Args:
-        repo (pygit2.Repository):
-            The repository object to search in.
-        expected_file_name (str):
-            The base file name to check (e.g., "readme").
-        check_extension (bool):
-            Whether to check the extension of the file or not.
-        extensions (list[str]):
-            List of possible file extensions to check (e.g., [".md", ""]).
-
-    Returns:
-        bool:
-            True if the file exists, False otherwise.
-    """
-
-    # Gather a tree from the HEAD of the repo
-    tree = repo.revparse_single("HEAD").tree
-
-    # Normalize expected file name to lowercase for case-insensitive comparison
-    expected_file_name = expected_file_name.lower()
-
-    for entry in tree:
-        # Normalize entry name to lowercase
-        entry_name = entry.name.lower()
-
-        # Check if the base file name matches with any allowed extension
-        if check_extension and any(
-            entry_name == f"{expected_file_name}{ext.lower()}" for ext in extensions
-        ):
-            return True
-
-        # Check whether the filename without an extension matches the expected file name
-        if not check_extension and entry_name.split(".", 1)[0] == expected_file_name:
-            return True
-
-    return False
-
-
-def is_citable(repo: pygit2.Repository) -> bool:
-    """
-    Check if the given repository is citable.
-
-    A repository is considered citable if it contains a CITATION.cff or CITATION.bib
-    file, or if the README.md file contains a citation section indicated by "## Citation"
-    or "## Citing".
-
-    Args:
-        repo (pygit2.Repository): The repository to check for citation files.
-
-    Returns:
-        bool: True if the repository is citable, False otherwise.
-    """
-
-    # Check for a CITATION.cff or CITATION.bib file
-    if file_exists_in_repo(
-        repo=repo,
-        expected_file_name="citation",
-        check_extension=True,
-        extensions=[".cff", ".bib"],
-    ):
-        return True
-
-    # Look for a README.md file and read its content
-    if (
-        file_content := read_file(
-            repo=repo, filepath="readme.md", case_insensitive=True
-        )
-    ) is not None:
-        # Check for an H2 heading indicating a citation section
-        if any(
-            check_string in file_content
-            for check_string in [
-                # markdown sub-headers
-                "## Citation",
-                "## Citing",
-                "## Cite",
-                "## How to cite",
-                # RST sub-headers
-                "Citation\n--------",
-                "Citing\n------",
-                "Cite\n----",
-                "How to cite\n-----------",
-                # DOI shield
-                "[![DOI](https://img.shields.io/badge/DOI",
-            ]
-        ):
-            return True
-
-    return False
-
-
-def default_branch_is_not_master(repo: pygit2.Repository) -> bool:
-    """
-    Checks if the default branch of the specified
-    repository is "master".
-
-    Args:
-        repo (Repository):
-            A pygit2.Repository object representing the Git repository.
-
-    Returns:
-        bool:
-            True if the default branch is "master", False otherwise.
-    """
-    # Access the "refs/remotes/origin/HEAD" reference to find the default branch
-    try:
-        # check whether remote head and remote master are the same
-        return (
-            repo.references.get("refs/remotes/origin/HEAD").target
-            == repo.references.get("refs/remotes/origin/master").target
-        )
-
-    except AttributeError:
-        # If "refs/remotes/origin/HEAD" or "refs/remotes/origin/master" doesn't exist,
-        # fall back to the local HEAD check
-        return repo.head.shorthand != "master"
-
-
 def days_of_development(repo: pygit2.Repository) -> float:
     """
 
@@ -246,114 +125,6 @@ def days_of_development(repo: pygit2.Repository) -> float:
 
     # Return the average commits per day
     return total_days
-
-
-def includes_common_docs(repo: pygit2.Repository) -> bool:
-    """
-    Check whether the repo includes common documentation files and directories
-    associated with building docsites.
-
-    Args:
-        repo (pygit2.Repository):
-            The repository object.
-
-    Returns:
-        bool:
-            True if any common documentation files
-            are found, False otherwise.
-    """
-    # List of common documentation file paths to check for
-    common_docs_paths = [
-        "docs/mkdocs.yml",
-        "docs/conf.py",
-        "docs/index.md",
-        "docs/index.rst",
-        "docs/index.html",
-        "docs/readme.md",
-        "docs/source/readme.md",
-        "docs/source/index.rst",
-        "docs/source/index.md",
-        "docs/src/readme.md",
-        "docs/src/index.rst",
-        "docs/src/index.md",
-    ]
-
-    # Check each documentation path using the find_file function
-    for doc_path in common_docs_paths:
-        if find_file(repo=repo, filepath=doc_path) is not None:
-            return True  # Return True as soon as we find any of the files
-
-    # otherwise return false as we didn't find documentation
-    return False
-
-
-def count_unique_contributors(
-    repo: pygit2.Repository, since: Optional[datetime] = None
-) -> int:
-    """
-    Counts the number of unique contributors to a repository.
-
-    If a `since` datetime is provided, counts contributors
-    who made commits after the specified datetime.
-    Otherwise, counts all contributors.
-
-    Args:
-        repo (pygit2.Repository):
-            The repository to analyze.
-        since (Optional[datetime]):
-            The cutoff datetime. Only contributions after
-            this datetime are counted. If None, all
-            contributions are considered.
-
-    Returns:
-        int:
-            The number of unique contributors.
-    """
-    since_timestamp = since.timestamp() if since else 0
-    contributors = {
-        commit.author.email
-        for commit in repo.walk(repo.head.target, pygit2.GIT_SORT_TIME)
-        if commit.commit_time > since_timestamp
-    }
-    return len(contributors)
-
-
-def count_repo_tags(repo: pygit2.Repository, since: Optional[datetime] = None) -> int:
-    """
-    Counts the number of tags in a pygit2 repository.
-
-    If a `since` datetime is provided, counts only tags associated
-    with commits made after the specified datetime. Otherwise,
-    counts all tags in the repository.
-
-    Args:
-        repo (pygit2.Repository):
-            The repository to analyze.
-        since (Optional[datetime]):
-            The cutoff datetime. Only tags for commits after
-            this datetime are counted. If None, all tags are counted.
-
-    Returns:
-        int:
-            The number of tags in the repository that meet the criteria.
-    """
-    since_timestamp = since.timestamp() if since else 0
-
-    count = 0
-    for ref in repo.references:
-        if ref.startswith("refs/tags/"):
-            tag = repo.lookup_reference(ref)
-            target_commit = repo[tag.target]
-
-            # Consider lightweight or annotated tags
-            if target_commit.type == pygit2.GIT_OBJECT_TAG:
-                target_commit = repo[target_commit.target]
-
-            # Check commit timestamp against `since`
-            if target_commit.commit_time > since_timestamp:
-                count += 1
-
-    return count
 
 
 def compute_repo_data(repo_path: str) -> None:
@@ -707,78 +478,6 @@ def _get_almanack_version() -> str:
         return almanack.__version__
 
 
-def get_api_data(
-    api_endpoint: str = "https://repos.ecosyste.ms/api/v1/repositories/lookup",
-    params: Optional[Dict[str, str]] = None,
-) -> dict:
-    """
-    Get data from an API based on the remote URL, with retry logic for GitHub rate limiting.
-
-    Args:
-        api_endpoint (str):
-            The HTTP API endpoint to use for the request.
-        params (Optional[Dict[str, str]])
-             Additional query parameters to include in the GET request.
-
-    Returns:
-        dict: The JSON response from the API as a dictionary.
-
-    Raises:
-        requests.RequestException: If the API call fails for reasons other than rate limiting.
-    """
-    if params is None:
-        params = {}
-
-    max_retries = 100  # Number of attempts for rate limit errors
-    base_backoff = 5  # Base backoff time in seconds
-
-    for attempt in range(1, max_retries + 1):
-        try:
-            # Perform the GET request with query parameters
-            response = requests.get(
-                api_endpoint,
-                headers={"accept": "application/json"},
-                params=params,
-                timeout=300,
-            )
-
-            # Raise an exception for HTTP errors
-            response.raise_for_status()
-
-            # Parse and return the JSON response
-            return response.json()
-
-        except requests.HTTPError as httpe:
-            # Check for rate limit error (403 with a rate limit header)
-            if (
-                response.status_code == 403  # noqa: PLR2004
-                and "X-RateLimit-Remaining" in response.headers
-            ):
-                if attempt < max_retries:
-                    # Calculate backoff time multiplied by attempt number
-                    # (linear growth)
-                    backoff = base_backoff * (attempt - 1)
-                    LOGGER.warning(
-                        f"Rate limit exceeded (attempt {attempt}/{max_retries}). "
-                        f"Retrying in {backoff} seconds..."
-                    )
-                    time.sleep(backoff)
-                else:
-                    LOGGER.warning("Rate limit exceeded. All retry attempts exhausted.")
-                    return {}
-            else:
-                # Raise other HTTP errors immediately
-                LOGGER.warning(f"Unexpected HTTP error: {httpe}")
-                return {}
-        except requests.RequestException as reqe:
-            # Raise other non-HTTP exceptions immediately
-            LOGGER.warning(f"Unexpected request error: {reqe}")
-            return {}
-
-    LOGGER.warning("All retries failed. Returning an empty response.")
-    return {}  # Default return in case all retries fail
-
-
 def get_github_build_metrics(
     repo_url: str,
     branch: str = "main",
@@ -978,208 +677,3 @@ def parse_python_coverage_data(
     # No recognized coverage files found
     LOGGER.warning("No coverage.py data found in the repository.")
     return {}
-
-
-def get_ecosystems_package_metrics(repo_url: str) -> Dict[str, Any]:
-    """
-    Fetches package data from the ecosystem API and calculates metrics
-    about the number of unique ecosystems, total version counts,
-    and the list of ecosystem names.
-
-    Args:
-        repo_url (str):
-            The repository URL of the package to query.
-
-    Returns:
-        Dict[str, Any]:
-            A dictionary containing information about packages
-            related to the repository.
-    """
-
-    if repo_url is None:
-        LOGGER.warning(
-            "Did not receive a valid repository URL and unable to gather package metrics."
-        )
-        return {}
-
-    # normalize http to https if necessary
-    if repo_url.startswith("http://"):
-        LOGGER.warning(
-            "Received `http://` repository URL for package metrics search. Normalizing to use `https://`."
-        )
-        repo_url = repo_url.replace("http://", "https://")
-
-    # normalize git@ ssh links to https if necessary
-    if repo_url.startswith("git@"):
-        LOGGER.warning(
-            "Received `git@` repository URL for package metrics search. Normalizing to use `https://`."
-        )
-        domain, path = repo_url[4:].split(":", 1)
-        repo_url = f"https://{domain}/{path}".removesuffix(".git")
-
-    # perform package srequest
-    package_data = get_api_data(
-        api_endpoint="https://packages.ecosyste.ms/api/v1/packages/lookup",
-        params={"repository_url": repo_url},
-    )
-
-    # Initialize counters
-    ecosystems = set()
-    total_versions = 0
-
-    for entry in package_data:
-        # count ecosystems
-        if "ecosystem" in entry:
-            ecosystems.add(entry["ecosystem"])
-
-        # sum versions
-        total_versions += entry.get("versions_count", 0)
-
-    return {
-        "ecosystems_count": len(ecosystems),
-        "versions_count": total_versions,
-        "ecosystems_names": sorted(ecosystems),
-    }
-
-
-def detect_social_media_links(content: str) -> Dict[str, List[str]]:
-    """
-    Analyzes README.md content to identify social media links.
-
-    Args:
-        readme_content (str):
-            The content of the README.md file as a string.
-
-    Returns:
-        Dict[str, List[str]]:
-            A dictionary containing social media details
-            discovered from readme.md content.
-    """
-    # Define patterns for social media links
-    social_media_patterns = {
-        "Twitter": r"https?://(?:www\.)?twitter\.com/[\w]+",
-        "LinkedIn": r"https?://(?:www\.)?linkedin\.com/(?:in|company)/[\w-]+",
-        "YouTube": r"https?://(?:www\.)?youtube\.com/(?:channel|c|user)/[\w-]+",
-        "Facebook": r"https?://(?:www\.)?facebook\.com/[\w.-]+",
-        "Instagram": r"https?://(?:www\.)?instagram\.com/[\w.-]+",
-        "TikTok": r"https?://(?:www\.)?tiktok\.com/@[\w.-]+",
-        "Discord": r"https?://(?:www\.)?discord(?:\.gg|\.com/invite)/[\w-]+",
-        "Slack": r"https?://[\w.-]+\.slack\.com",
-        "Gitter": r"https?://gitter\.im/[\w/-]+",
-        "Telegram": r"https?://(?:www\.)?t\.me/[\w-]+",
-        "Mastodon": r"https?://[\w.-]+/users/[\w-]+",
-        "Threads": r"https?://(?:www\.)?threads\.net/[\w.-]+",
-        "Bluesky": r"https?://(?:www\.)?bsky\.app/profile/[\w.-]+",
-    }
-
-    # Initialize results
-    found_platforms = set()
-
-    # Search for social media links
-    for platform, pattern in social_media_patterns.items():
-        if re.search(pattern, content, re.IGNORECASE):
-            found_platforms.add(platform)
-
-    return {
-        "social_media_platforms": sorted(found_platforms),
-        "social_media_platforms_count": len(found_platforms),
-    }
-
-
-def find_doi_citation_data(repo: pygit2.Repository) -> Dict[str, Any]:
-    """
-    Find and validate DOI information from a CITATION.cff
-    file in a repository.
-
-    This function searches for a `CITATION.cff` file in the provided repository,
-    extracts the DOI (if available), validates its format, checks its
-    resolvability via HTTP, and performs an exact DOI lookup on the OpenAlex API.
-
-    Args:
-        repo (pygit2.Repository):
-            The repository object to search for the CITATION.cff file.
-
-    Returns:
-        Dict[str, Any]:
-            A dictionary containing DOI-related information and metadata.
-    """
-
-    result = {
-        "doi": None,
-        "valid_format_doi": None,
-        "https_resolvable_doi": None,
-        "publication_date": None,
-        "cited_by_count": None,
-    }
-
-    # Find the CITATION.cff file
-    if (citationcff_file := find_file(repo=repo, filepath="CITATION.cff")) is None:
-        LOGGER.warning("No CITATION.cff file discovered.")
-        return result
-
-    try:
-        # Read and parse the CITATION.cff file
-        citation_data = yaml.safe_load(read_file(repo=repo, entry=citationcff_file))
-
-        # Extract DOI from 'doi' or 'identifiers' field
-        if "doi" in citation_data.keys():
-            result["doi"] = citation_data.get("doi", None)
-        elif "identifiers" in citation_data.keys():
-            result["doi"] = next(
-                (
-                    identifier["value"]
-                    for identifier in citation_data.get("identifiers", [])
-                    if identifier.get("type") == "doi"
-                ),
-                None,
-            )
-    except yaml.YAMLError as e:
-        LOGGER.warning(f"Error reading YAML: {e}")
-
-    if result["doi"]:
-        # Validate the DOI format
-        result["valid_format_doi"] = bool(
-            re.match(r"^10\.\d{4,9}/[-._;()/:A-Za-z0-9]+$", result["doi"])
-        )
-        if result["valid_format_doi"]:
-            try:
-                # Check DOI resolvability via HTTPS
-                if (
-                    requests.head(
-                        f"https://doi.org/{result['doi']}",
-                        allow_redirects=True,
-                        timeout=30,
-                    ).status_code
-                    == 200  # noqa: PLR2004
-                ):
-                    result["https_resolvable_doi"] = True
-                else:
-                    LOGGER.warning(
-                        f"DOI does not resolve properly: https://doi.org/{result['doi']}"
-                    )
-                    result["https_resolvable_doi"] = False
-
-            except requests.RequestException as e:
-                LOGGER.warning(f"Error resolving DOI: {e}")
-                result["https_resolvable_doi"] = False
-
-            # Perform exact DOI lookup on OpenAlex
-            try:
-                openalex_result = get_api_data(
-                    api_endpoint=f"https://api.openalex.org/works/doi:{result['doi']}"
-                )
-                publication_date = openalex_result.get("publication_date", None)
-                result.update(
-                    {
-                        "publication_date": (
-                            datetime.strptime(publication_date, "%Y-%m-%d")
-                            if publication_date is not None
-                            else None
-                        ),
-                        "cited_by_count": openalex_result.get("cited_by_count", None),
-                    }
-                )
-            except requests.RequestException as e:
-                LOGGER.warning(f"Error during OpenAlex exact DOI lookup: {e}")
-
-    return result
