+++ src/book/software-forest/software-forest.md
@@ -11,7 +11,7 @@ The software forest is a multidimensional garden, enriched by connections and mi
 Both the practice and artifacts of software development can seem daunting, like an unfamiliar forest full of mystery and surprises.
 However, through software gardening, we see these forests as opportunities for balanced cultivation.
 
-> "Called Three-Dimensional Forestry or Forest Farming, … the three 'dimensions' of \[Toyohiko Kagawa's\] 3-D system were the trees as conservers of the soil and suppliers of food and the livestock which benefited from them." {cite:p}`hart_forest_1996`, {cite:p}`kagawa_fujisaki_3d_farming`
+> "Called Three-Dimensional Forestry or Forest Farming, … the three 'dimensions' of [Toyohiko Kagawa's] 3-D system were the trees as conservers of the soil and suppliers of food and the livestock which benefited from them." {cite:p}`hart_forest_1996`, {cite:p}`kagawa_fujisaki_3d_farming`
 
 Forest gardening offers innovations from an ecological cultivation perspective by caring for trees, understory vegetation, soil, and other environmental factors.
 Forest gardens take time to create because their slow-growing nature requires patience and careful, long-term nurturing.
